Update WOPR_Display_V2.ino
moved configTime() call into a separate function called syncTimeWithNTP() and added a hard-coded 30 second timeout. this was done because i was getting some time sync errors because the sync was taking a little longer than the code allowed for. waiting for sync to complete is important if i'm starting up straight into clock mode.